Renamed usePassPropsToChildren to passPropsToChildren
